scripts/requirements.txt
requests
python-dotenv
tldextract
